LayoutTests/webanimations/threaded-animations/view-timeline-subject-metrics-change-reflected-on-remote-progress-based-timeline.html

<!DOCTYPE html>
<body>
<style>

body {
    height: 2000px;
}

.subject {
    position: absolute;
    top: calc(100vw - 50px);
    width: 100px;
    height: 100px;
    background-color: black;
}

</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/ui-helper.js"></script>
<script src="../../imported/w3c/web-platform-tests/web-animations/testcommon.js"></script>
<script src="threaded-animations-utils.js"></script>

<script>

const make_subject = t => {
    const subject = createDiv(t);
    subject.className = "subject";
    return subject;
};

promise_test(async t => {
    const maxScrollTop = document.documentElement.scrollHeight - window.innerHeight;
    window.scrollTo(0, maxScrollTop / 2);

    const subject = make_subject(t);
    const timeline = new ViewTimeline({ subject });
    const animation = subject.animate({ translate: '100px' }, { timeline });
    await animationAcceleration(animation);
    const initialTimeline = await UIHelper.remoteTimeline(timeline);
    assert_not_equals(initialTimeline, undefined, "Remote timeline was created.");

    subject.style.height = "200px";
    await threadedAnimationsCommit();
    const updatedTimeline = await UIHelper.remoteTimeline(timeline);
    assert_not_equals(updatedTimeline.currentTime, initialTimeline.currentTime, "Changing a timeline's subject size changes the remote timeline's current time.");
}, "A remote progress-based timeline is updated when a view timeline's subject changes metrics.");

</script>
</body>
